Fix build errors under 64bit compiler

+++ ddbc/drivers/sha1.d
@@ -205,7 +205,7 @@ public:
       return rv;
    }
 
-   bool input(const(ubyte)* message_array, uint length)
+   bool input(const(ubyte)* message_array, ulong length)
    {
       if (!length)
          return true;
